+++ cpa-repo/src/main/kotlin/no/nav/emottak/cpa/validation/CPAValidator.kt
@@ -5,7 +5,7 @@ import no.nav.emottak.cpa.getPartyInfoByTypeAndID
 import no.nav.emottak.message.ebxml.EbXMLConstants.ACKNOWLEDGMENT_ACTION
 import no.nav.emottak.message.ebxml.EbXMLConstants.EBMS_SERVICE_URI
 import no.nav.emottak.message.ebxml.EbXMLConstants.MESSAGE_ERROR_ACTION
-import no.nav.emottak.message.model.Addressing
+import no.nav.emottak.utils.common.model.Addressing
 import no.nav.emottak.message.model.ValidationRequest
 import org.oasis_open.committees.ebxml_cppa.schema.cpp_cpa_2_0.CollaborationProtocolAgreement
 import org.oasis_open.committees.ebxml_cppa.schema.cpp_cpa_2_0.PartyInfo
