+++ cpa-repo/src/test/kotlin/no/nav/emottak/cpa/TestUtil.kt
@@ -1,10 +1,10 @@
 package no.nav.emottak.cpa
 
-import no.nav.emottak.message.model.Addressing
+import no.nav.emottak.utils.common.model.Addressing
 import no.nav.emottak.message.model.Direction.IN
 import no.nav.emottak.message.model.Header
-import no.nav.emottak.message.model.Party
-import no.nav.emottak.message.model.PartyId
+import no.nav.emottak.utils.common.model.Party
+import no.nav.emottak.utils.common.model.PartyId
 import no.nav.emottak.message.model.ValidationRequest
 import no.nav.emottak.util.createCRLFile
 import no.nav.emottak.util.createX509Certificate
