+++ ebms-async/src/test/kotlin/no/nav/emottak/ebms/async/persistence/EventsRepositoryTest.kt
@@ -4,10 +4,10 @@ import no.nav.emottak.ebms.async.ebmsPostgres
 import no.nav.emottak.ebms.async.persistence.repository.EbmsMessageDetailsRepository
 import no.nav.emottak.ebms.async.persistence.repository.EventsRepository
 import no.nav.emottak.ebms.async.testConfiguration
-import no.nav.emottak.message.model.Addressing
+import no.nav.emottak.utils.common.model.Addressing
 import no.nav.emottak.message.model.Event
-import no.nav.emottak.message.model.Party
-import no.nav.emottak.message.model.PartyId
+import no.nav.emottak.utils.common.model.Party
+import no.nav.emottak.utils.common.model.PartyId
 import no.nav.emottak.message.model.Payload
 import no.nav.emottak.message.model.PayloadMessage
 import no.nav.emottak.message.model.toEbmsMessageDetails
