+++ ebms-provider/src/main/kotlin/no/nav/emottak/ebms/HttpClients.kt
@@ -25,8 +25,8 @@ import no.nav.emottak.message.model.MessagingCharacteristicsRequest
 import no.nav.emottak.message.model.MessagingCharacteristicsResponse
 import no.nav.emottak.message.model.PayloadRequest
 import no.nav.emottak.message.model.PayloadResponse
-import no.nav.emottak.message.model.SendInRequest
-import no.nav.emottak.message.model.SendInResponse
+import no.nav.emottak.utils.common.model.SendInRequest
+import no.nav.emottak.utils.common.model.SendInResponse
 import no.nav.emottak.message.model.ValidationRequest
 import no.nav.emottak.message.model.ValidationResult
 import no.nav.emottak.utils.common.model.DuplicateCheckRequest
