+++ ebms-provider/src/main/kotlin/no/nav/emottak/ebms/processing/ProcessingService.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.withContext
 import no.nav.emottak.ebms.PayloadProcessingClient
 import no.nav.emottak.ebms.logger
 import no.nav.emottak.melding.feil.EbmsException
-import no.nav.emottak.message.model.Addressing
+import no.nav.emottak.utils.common.model.Addressing
 import no.nav.emottak.message.model.Direction
 import no.nav.emottak.message.model.Payload
 import no.nav.emottak.message.model.PayloadMessage
