+++ ebms-provider/src/main/kotlin/no/nav/emottak/ebms/util/EventRegistration.kt
@@ -2,7 +2,7 @@ package no.nav.emottak.ebms.util
 
 import no.nav.emottak.ebms.log
 import no.nav.emottak.message.model.EbMSDocument
-import no.nav.emottak.message.model.PartyId
+import no.nav.emottak.utils.common.model.PartyId
 import no.nav.emottak.utils.common.parseOrGenerateUuid
 import no.nav.emottak.utils.kafka.model.EbmsMessageDetail
 import no.nav.emottak.utils.kafka.model.Event
