+++ ebms-provider/src/test/kotlin/no/nav/emottak/ebms/EbmsRouteSyncIT.kt
@@ -21,17 +21,17 @@ import no.nav.emottak.ebms.model.signer
 import no.nav.emottak.ebms.sendin.SendInService
 import no.nav.emottak.ebms.validation.CPAValidationService
 import no.nav.emottak.ebms.validation.MimeHeaders
-import no.nav.emottak.message.model.Addressing
+import no.nav.emottak.utils.common.model.Addressing
 import no.nav.emottak.message.model.Direction
 import no.nav.emottak.message.model.EbMSDocument
-import no.nav.emottak.message.model.EbmsProcessing
+import no.nav.emottak.utils.common.model.EbmsProcessing
 import no.nav.emottak.message.model.ErrorCode
 import no.nav.emottak.message.model.Feil
-import no.nav.emottak.message.model.Party
-import no.nav.emottak.message.model.PartyId
+import no.nav.emottak.utils.common.model.Party
+import no.nav.emottak.utils.common.model.PartyId
 import no.nav.emottak.message.model.PayloadMessage
 import no.nav.emottak.message.model.PayloadProcessing
-import no.nav.emottak.message.model.SendInResponse
+import no.nav.emottak.utils.common.model.SendInResponse
 import no.nav.emottak.message.model.SignatureDetails
 import no.nav.emottak.message.model.ValidationResult
 import no.nav.emottak.util.decodeBase64
