+++ ebxml-processing-model/src/main/kotlin/no/nav/emottak/message/ebxml/MessageHeaderExt.kt
@@ -3,9 +3,9 @@
  */
 package no.nav.emottak.message.ebxml
 
-import no.nav.emottak.message.model.Addressing
-import no.nav.emottak.message.model.Party
-import no.nav.emottak.message.model.PartyId
+import no.nav.emottak.utils.common.model.Addressing
+import no.nav.emottak.utils.common.model.Party
+import no.nav.emottak.utils.common.model.PartyId
 import org.oasis_open.committees.ebxml_msg.schema.msg_header_2_0.From
 import org.oasis_open.committees.ebxml_msg.schema.msg_header_2_0.MessageData
 import org.oasis_open.committees.ebxml_msg.schema.msg_header_2_0.MessageHeader
