+++ ebxml-processing-model/src/main/kotlin/no/nav/emottak/message/model/MessageError.kt
@@ -2,6 +2,9 @@ package no.nav.emottak.message.model
 
 import no.nav.emottak.message.ebxml.EbXMLConstants
 import no.nav.emottak.message.xml.xmlMarshaller
+import no.nav.emottak.utils.common.model.Addressing
+import org.oasis_open.committees.ebxml_msg.schema.msg_header_2_0.ErrorList
+import org.oasis_open.committees.ebxml_msg.schema.msg_header_2_0.SeverityType
 import org.w3c.dom.Document
 import org.xmlsoap.schemas.soap.envelope.Body
 import org.xmlsoap.schemas.soap.envelope.ObjectFactory
@@ -33,4 +36,32 @@ data class MessageError(
             EbMSDocument(requestId, it, emptyList())
         }
     }
+}
+
+@JvmName("asErrorList")
+fun List<Feil>.asErrorList(): ErrorList {
+    if (this.isEmpty()) {
+        throw IllegalArgumentException("(4.2.3 Kan ikke opprette ErrorList uten errors")
+    }
+
+    return this.map {
+        it.code.createEbxmlError(it.descriptionText, if (it.severity != null) SeverityType.fromValue(it.severity) else null)
+    }.asErrorList()
+}
+
+@JvmName("toErrorList")
+fun List<org.oasis_open.committees.ebxml_msg.schema.msg_header_2_0.Error>.asErrorList(): ErrorList {
+    if (this.isEmpty()) {
+        throw IllegalArgumentException("(4.2.3 Kan ikke opprette ErrorList uten errors")
+    }
+    val errorList = ErrorList()
+    errorList.error.addAll(this)
+    errorList.version = "2.0"
+    // errorList.any // Unused?
+    errorList.id // "May be used for error tracking"
+    errorList.highestSeverity = this.sortedBy {
+        it.severity == SeverityType.ERROR
+    }.first().severity
+    errorList.isMustUnderstand = true // Alltid
+    return errorList
 }
