+++ ebxml-processing-model/src/main/kotlin/no/nav/emottak/message/util/MimeUtil.kt
@@ -2,6 +2,7 @@ package no.nav.emottak.message.util
 
 import kotlin.uuid.ExperimentalUuidApi
 import kotlin.uuid.Uuid
+import no.nav.emottak.utils.environment.getEnvVar
 
 @OptIn(ExperimentalUuidApi::class)
 fun createUniqueMimeMessageId() = "${Uuid.random()}@$hostName"
@@ -11,7 +12,3 @@ private val hostName = when (getEnvVar("NAIS_CLUSTER_NAME", "local")) {
     "dev-fss" -> "dev.ebms.nav.no"
     else -> "local"
 }
-
-// For å slippe avhengighet til emottak-utils (kopiert fra EnvUtils.kt):
-private fun getEnvVar(varName: String, defaultValue: String? = null) =
-    System.getProperty(varName) ?: System.getenv(varName) ?: defaultValue ?: throw RuntimeException("Environment: Missing required variable \"$varName\"")
