--- felles/src/main/kotlin/no/nav/emottak/constants/LogIndex.kt
@@ -1,15 +1,0 @@
-package no.nav.emottak.constants
-
-object LogIndex {
-    const val MESSAGE_ID = "messageId"
-    const val CONVERSATION_ID = "conversationId"
-    const val FROM_PARTY = "avsenderId"
-    const val TO_PARTY = "mottakerId"
-    const val TO_ROLE = "role"
-    const val FROM_ROLE = "fromRole"
-    const val SERVICE = "service"
-    const val CPA_ID = "cpaId"
-    const val ACTION = "action"
-    const val X_MAILER = "systemkilde"
-    const val X_REQUEST_ID = "requestId"
-}
