+++ felles/src/main/kotlin/no/nav/emottak/util/LoggingUtil.kt
@@ -3,22 +3,22 @@ package no.nav.emottak.util
 import io.ktor.http.Headers
 import net.logstash.logback.marker.LogstashMarker
 import net.logstash.logback.marker.Markers
-import no.nav.emottak.constants.LogIndex.ACTION
-import no.nav.emottak.constants.LogIndex.CONVERSATION_ID
-import no.nav.emottak.constants.LogIndex.CPA_ID
-import no.nav.emottak.constants.LogIndex.FROM_PARTY
-import no.nav.emottak.constants.LogIndex.FROM_ROLE
-import no.nav.emottak.constants.LogIndex.MESSAGE_ID
-import no.nav.emottak.constants.LogIndex.SERVICE
-import no.nav.emottak.constants.LogIndex.TO_PARTY
-import no.nav.emottak.constants.LogIndex.TO_ROLE
-import no.nav.emottak.constants.LogIndex.X_MAILER
-import no.nav.emottak.constants.LogIndex.X_REQUEST_ID
+import no.nav.emottak.utils.common.constants.LogFields.ACTION
+import no.nav.emottak.utils.common.constants.LogFields.CONVERSATION_ID
+import no.nav.emottak.utils.common.constants.LogFields.CPA_ID
+import no.nav.emottak.utils.common.constants.LogFields.FROM_PARTY
+import no.nav.emottak.utils.common.constants.LogFields.FROM_ROLE
+import no.nav.emottak.utils.common.constants.LogFields.MESSAGE_ID
+import no.nav.emottak.utils.common.constants.LogFields.SERVICE
+import no.nav.emottak.utils.common.constants.LogFields.TO_PARTY
+import no.nav.emottak.utils.common.constants.LogFields.TO_ROLE
+import no.nav.emottak.utils.common.constants.LogFields.X_MAILER
+import no.nav.emottak.utils.common.constants.LogFields.X_REQUEST_ID
 import no.nav.emottak.constants.SMTPHeaders
 import no.nav.emottak.message.model.EbmsMessage
 import no.nav.emottak.message.model.Header
 import no.nav.emottak.message.model.PayloadRequest
-import no.nav.emottak.message.model.SendInRequest
+import no.nav.emottak.utils.common.model.SendInRequest
 import no.nav.emottak.message.model.SignatureDetailsRequest
 import no.nav.emottak.message.model.ValidationRequest
 import org.oasis_open.committees.ebxml_msg.schema.msg_header_2_0.MessageHeader
